Add Apache Parquet support for TensorFlow Dataset
Apache Parquet is a widely used columnar storage format available in
the Hadoop ecosystem. Its popularity spans across big data community
with usage in many production environments. For example, most of our
big data projects use Parquet format for storage on S3 processed by
Spark/EMR. We started transition some of projects to TensorFlow.
However, as there was no Parquet support in TensorFlow, we have to
transform our existing Parquet data to formats accepted by TensorFlow first.

This PR is a preliminary attempt to add Apache Parquet support for
TensorFlow Dataset. It may not cover all the use cases though it could
be served as a starting point for further improvement in the future.

The ParquetDataset depends on parquet-cpp (Apache) project as well
as other dependencies. The ParquetDataset only builds on Linux at the
moment. This PR also adds the option in `./configure` so
that those dependencies could be skipped.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>