Add write_kafka ops for streaming output
This commit adds a very early implementation of write_kafka ops
so that we could do a streaming output.

The basic idea of write_kafka is to implement in C++ WriteKafkaOp
which takes a message as input, write the message to a Kafka server,
and returns the original message. The return of original message
is on purpose so that write_kafka is not a terminal op.
This helps to do further processing.

So far most of the tensorflow-io ops are input, while the output
has been mentioned in discussions. This implementation could be
a start for future implementation of output ops.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>