Fix tiff lib and remove stub ffmpeg libraries for macOS support
The video ops used to be linked to a stub (empty) library
so that in compile time it is possible to build tensorflow-io
without actually providing ffmpeg dynamic library.
This causes issues in bazel's sandboxed mode because sandboxed
one need to know all files in advance. So far we have been
using standalone mode to address this issue. It is still
desirable to use sandbox mode as there might be some unintended
effects. For example, when we compile tiff we assume
jpeg is not supported, but in Linux it is almost certain it will
include a header from system library. Which is one of the
reasons causeing MacOS to fail (include path different from linux).

In short, we should use sandboxed mode for Bazel.

This PR fixes the TIFF issue, and also uses dlopen(RTLD_GLOBAL)
to resolve ffmpeg symbols in run time, effectively avoids
the stub (empty) libary we have.

This PR is part of the effort to spport macOS for tensorflow-io.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>