Add CIFAR dataset support with compression
This PR adds the CIFAR data format support. Unlike MNIST data format,
which has been used in quite a few other places like:
- MNIST
- Fashion MNIST
- Kuzushiji-MNIST
- EMNIST
, the CIFAR is not widely used (at least not widely seen) other
than its original CIFAR10/CIFAR100 format. However, CIFAR is
a good candidate for us to look into the archive/compression
which will be used for other places. CIFAR is limited
so the impact of this implementation is limited. We could
easily revisit and rework on CIFAR once we have a better
understanding of compression/archive format.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>