Add ImageInputSequence class for metadata information with a list of images
This PR adds an ImageInputSequence class so that metadata information
of a list of images (one image could have multiple frames) is stored
as a resource. The meta data information could later be
obtained without re-compute (as the information are static).

At the moment two methods are added: `count()` which returns
the number of frames, and `type()` which returns the fourcc
file type of each frame.

Additional meta data information could be added later.

This PR is inspired by 109.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>