Add batch support for MNISTDataset
This fix add batch support for MNISTDataset. Originally,
Dataset does not have a built in batch support at the time
of Dataset creation. The batch is typically specified
with `batch(n)` call. This helps with low mmeory footprint
but may not be performance when working with small records
data where each record may only have several bytes.

This fix adds the option to specifiy the batch at creation,
and allows to read a batch (e.g, 1000) to improve I/O
throughput.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>