Add GRPCDataset to allow pulling data from a gRPC server
This fix tries to have a way to pull data from a gRPC server,
as long as a gRPC server implement the protocol in endpoint.proto.

The main purpose is to allow reading partial data from numpy array
in memory.

In the from_numpy method, the following is done:
- Create a gRPC server and exposes ReadRecord endpoint.
- Start the gRPC server on a random port in local host.
- Pass endpoint (<localhost:port>) to GRPCDataset
- GRPCDataset will pull the data from gRPC server.

Note from_numpy is just a facility method to setup a gRPC
server. In theory, a GRPC server could be created by any other
process and by any language. Only the endpoint information is needed
to create GRPCDataset.

In case numpy array is huge, then this method could be helpful
as it is not required to save numpy file, and load back by tf.data,
or pass the whole numpy data into a tensor.

This could open up doors for other languages as well. For example,
in case of R, we could setup a grpc server with R, and pass R dataframes
in memory and allows GRPCDataset to pull a chunk of the data at a time.

Note we create gRPC server locally, but it is possible to expose gRPC
server remotely so that the workload could be distributed.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>