Add save method for TextDataset
This is an experimental implementation of the `save` (write)
method for TextDataset in order to allow saving
the content of Dataset into files.

TextDataset itself is straightforward so it may not be that useful.
However, it setup an example for saving to other future format (e.g., Parquet/Arrow/etc).

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>