Rework on PrometheusDataset
This is part of the effort to reduce the dedicated C++
implementation of Dataset and replace with primitive ops
that could be used both with tf.data, and with Tensor.

There are some room for enhancement, for example,
a timestamp could be passed to read_prometheus
and each call will only read a small slice of the data
to tensor. Will have follow up PRs later to implement that.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>