Add list_feather_columns function in eager mode
This PR adds list_feather_columns function in eager mode,
so that it is possible to get the column name and spec
information for feather format.

This PR implements an `::arrow::io::RandomAccessFile` interface
so it is possible to read files through scheme file system,
e.g., s3, gcs, azfs, etc.

The `::arrow::io::RandomAccessFile` is the same as in Parquet PR 384
so they could be combined.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>