Add read_wav and list_wav_info support
This PR adds read_wav and list_wav_info support
so that it is possible to read wav file into a tensor.
Since WAV file is splittable, read_wav could take a
start and count parameter so that only a slice
of the wav file is read.

This is also part of the effort to rework on Dataset
to move to primitive ops.

See See 382 and 366 for related discussions.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>