Add read_archive and list_archive_entries support (#389)
* Add read_archive and read_archive_entries support

This PR is part of the discussion in 366 where we want
to reduce the C++ implementation by reusing primitive ops
for archive support.

The idea is to implement C++ kernel that could be used in general,
than adding additional C++ for each Dataset.

Summary:
1) read_archive_entries could probe for archive format and returns entry
   names in archive automatically.
2) read_archive can take the output of read_archive_entries, and returns
   string tensor as content of the entries in archive.
3) both read_archive_entries and read_archive support in graph and eager mode
   and could be wired to implement an "ArchiveDataset" if needed
   (see demo in tests/test_archive_eager.py)
4) combined with other PR 384 it is possible to read memory and decode
   parquet like file in memory, by utilizing `SizedRandomAccessFile`
   which is a subclass of tensorflow's `RandomAccessFile`.

Overall, we expect files compressed in archive not to be too big so it should
at least fit into CPU memory. While files such as HDF5 or parquet could be really
large, the compression normally happens at the file format level so additional
archive/compression outside with huge file (e.g. hundreds of GBs) are rarely heard
of. So, the implementation in this PR should be enough for normal need.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Fix macOS build failure

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Fix 1.14 build failure

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Rename to list_archive_entries and read_archive

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>