Expose top level tfio.Dataset as a superset of tf.data.Dataset
This PR expose top level tfio.Dataset as a superset of tf.data.Dataset.

In the past, we have been trying to place different Dataset implementations
into individual modules. For example, we have tensorflow_io.arrow.ArrowFeatherDataset
under `tensorflow_io.arrow`. As the project grow we already have too many modules and
it is getting harder to find the features we want.

This PR tries to expose a top level tfio.Dataset as a superset of tf.data.Dataset,
so that users could use tfio.Dataset the same way as tf.data.Dataset and
with additional functionalities.

To achive that first we have to expose all functions in tf.data.Dataset to tfio.Dataset.
Since tfio.Dataset is a subclass we also have to fix the return types so that
`tfio.Dataset.range(5)` will returns a subclass of `tfio.Dataset` (even though
we redirect tfio.Dataset.range(5) to tf.data.Dataset.range(5) which by nature
will return a subclass of `tf.data.Dataset`.)

This PR creates an wrapper of `_DatasetAdapter` and `dataset_decorator` so that
return type could be adjusted.

Also, all functions/methods of `tf.data.Dataset` are routed with `tfio.Dataset`
automatically. (We could decide to manually define those functions/methods instead,
if we feel it is much better).

In addition to create a class of tfio.Dataset, this PR also adds the `rebatch()`
as part of `tfio.Dataset`, so that user could use it in an natural way.
```
tfio.Dataset.range(5).batch(3).rebatch(5)
```

Further, I think we could also start exposing individual datasets as part of
the static method in `tfio.Dataset`. For example, we could imagine:
```
tfio.Dataset.from_arrow(filename)
```
which would be very similiar to:
```
tf.data.Dataset.from_tensors(...)
```

Note those static methods of `from_arrow` has not been added yet, we could discuss
what would be the best API to expose those.

Finally the way of using `inspect.getmembers` with `_DatasetAdapter` and `dataset_decorator`
in this PR may not be the best way. A more pythonic way might be to use metaclass.
However, as tf.data.Dataset is also a metaclass so things are complicated. This PR is
what I can think of so far but ideas or contributions to make the mechanism more elegant
are welcomed.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>