Expose tfio.IOTensor class and from_audio
This PR tries to expose a tfio.IOTensor which could
be applied to and io related data which are indexable
(`__getitem__` and `__len__`)

The idea is to bind `__getitem__` and `__len__` to
kernel ops in run time, so that is is not necessarily
to read everything in memory.

The first file format is the WAV file. With tfio.IOTensor
dtype and shape are exposed with `__getitem__` and `__len__`.

Further, a `rate` property has been exposed specifically
for Audio/WAV file which gives sample rate.

This tfio.IOTensor only works in eager mode.

In additional this PR also converts WavDataset to
use IOTensor (instead of direct C++ implementation).

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>