Add tfio.IOTensor.from_feather support
Note: this PR depends on PR 438.

Feather is a columnar file format that often seen with pandas.
This PR adds the indexing and slicing support to bring
Feather to parity with Parquet file format, by adding
tfio.IOTensor.from_feather support so that it is possible to access
feather through natual `__getitem__` operations.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>