Add support to use sklearn.preprocessing on top of IOTensor, and add `from_tensor` support
sklearn.preprocessing is widely used for data preprocessing so allowing users
to apply sklearn.preprocessing on top of IOTensor directly will greatly help
user.

The background was to try to build a model on series data with LSTM, for Prometheus
and Kafka. Normally applying LSTM needs
1) Preprocess/Normalize data, sklearn.preprocessing could be handy to avoid additional implementation.
2) A sliding window operation to create a shape with look back in place.

This PR will help the first case. A follow up PR will be for 2).

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>