Add window() operation support for IOTensor
This PR adds window operation support for IOTensor.

The main purpose of the `window` operation is to allow
processing time series data (such as from prometheus)
to be used in LSTM training. In LSTM training the data
are normally preprocessed as a tuple of `(feature(n-1), label(1))`
where `n-1` is the look back.

It seems that in tensorflow there is no direct window-like kernel ops.
This PR converts the data into an image of `[1, 1, len, 1]` and
use `tf.image.extract_patchs` to achieve similiar effect.

Only 1D data is supported at the moment but expand the window op
to generic IOTensor should be possible.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>