Add wrapper to use tf.keras.util.Sequence
There are multiple ways to pass data to tf.keras. While
dataset could be used, it assumes that dataset is always repeatable
which means multiple runs produce the same type of data.
But that is not necesssarily the case for streaming
where data is not repeatable.

For example, prometheus largely depends on the time to
generate the data so unless the timestamp is passes, multiple
runs will have different data produced.

To avoid confusion, this PR adds a wrapper to tf.keras.util.Sequence
so that data is guaranteed to be repeatable when data is passed to tf.keras.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>