Add IO related batch operations on list of objects
In TensorFlow, IO related operations are either through gfile
which is a python implementation, or through read_file/write_file
which is a graph implementation. What is missing is the
batch operations on list of object(files).

This PR adds IO related batch operations on list of objects,
so that it is possible to:
1) Read chunks of data from a list of files based on offset/length
2) Get the file sizes from a list of files.

The motivations of IO related batch operations could be many:
1) Many file types could be easily detected through the magic number
   so reading the first few bytes is enough.
2) Many operations already expect batches so reading files in batch
   will help reduce intermediate steps.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>