Add tfio.IOTensor.from_avro support
Avro is a columnar file format that naturally fits into a table/column data.
Avro file itself is not directly indexable. However, it is pseudo-indexable
as it consists of blocks with each blocks specifying file offset/size, and
count of items. So indexing coulb be done by small range iteration.

It would be desirable to make Avro indexable as it will be much more
convenient with increased flexibility.

This PR adds tfio.IOTensor.from_avro support so that it is possible to acess
avro data through natual __getitem__ operations.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>