Add tfio.IOTensor.from_avro support (#440)
* Add tfio.IOTensor.from_avro support

Avro is a columnar file format that naturally fits into a table/column data.
Avro file itself is not directly indexable. However, it is pseudo-indexable
as it consists of blocks with each blocks specifying file offset/size, and
count of items. So indexing coulb be done by small range iteration.

It would be desirable to make Avro indexable as it will be much more
convenient with increased flexibility.

This PR adds tfio.IOTensor.from_avro support so that it is possible to acess
avro data through natual __getitem__ operations.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Add a Partitions function to Avro, so that it is possible

to dynamically adjust the capacity of the chunk size when reading.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Rename to io_stream.h for consistency

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Remove the need to pass component, unless needed explicitly

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Move Partitions to a generic location and support dataset

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>