Expose __version__ in tensorflow_io
This PR exposes __version__ in tensorflow_io, as was discussed
in 482. The single source of truth is located in
tensorflow_io/core/python/ops/io_info.py

so that any change only happens in one place.

A test case has been added as well to guarantee the correctness of the __version__
string for package release.

This PR fixes 482.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>