Update decode_tiff_info to allow float data support
In 534 we disscussed the native data support for tiff formats.
In order to extract the native data type we first have to
correctly identify the data type (as opposed always do the RGBA 32 conversion
in existing code).

This PR checks additional tiff field so that different data types could be
probed.

This PR is related to 534.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>