Add initial GeoTIFF support
This PR adds the initial GeoTIFF support. GeoTIFF is a format extension
of original TIFF with georeferencing information embedded.

It is widely used in geological information. Examples includs
"Landsat" from Google Cloud public dataset:

https://cloud.google.com/storage/docs/public-datasets/landsat

While GEOTiff is an extension of TIFF, libtiff will not be able
to extract geo tags and libgeotiff is needed.

This PR is the first step to build libgeotiff and extract images.

This PR is not processing geo tags yet, which very much
important for the GeoTIFF file to be useful.
But this could be the first step and we could use follow up PRs to
see how to best present the tags in a natural way to allow training or inference.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>