Move bigtable.so to libtensorflow_io.so
This PR is part of the effort in resolving 489. In 489
there are several .so files that could not be imported by
python correctly. While continue to investigate the issue
it is worthwhile to combine multiple .so to libtensorflow_io.so
to avoid multiple copies of library anyway.

This PR merges bigtable.so to libtensorflow_io.so

Note since there is already a copy of GrpcStatusToTfStatus
this PR comment out the duplicated one.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>