Remove BUILD file from tensorflow_io/cloud
It looks like tensorflow_io/cloud is covered
by tensorflow_io/gcs, and tensorflow_io/bigquery. Also
tensorflow_io/cloud is not truly working as it depends on
ReaderBase which seems to be deprecated in tensorflow's core rep.

This PR removes BUILD file tensorflow_io/cloud so that
it will not be build and picked up by wheel package

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>