Move bigtable.so to libtensorflow_io.so (#588)
* Move bigtable.so to libtensorflow_io.so

This PR is part of the effort in resolving 489. In 489
there are several .so files that could not be imported by
python correctly. While continue to investigate the issue
it is worthwhile to combine multiple .so to libtensorflow_io.so
to avoid multiple copies of library anyway.

This PR merges bigtable.so to libtensorflow_io.so

Note since there is already a copy of GrpcStatusToTfStatus
this PR comment out the duplicated one.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Remove BUILD file from tensorflow_io/cloud

It looks like tensorflow_io/cloud is covered
by tensorflow_io/gcs, and tensorflow_io/bigquery. Also
tensorflow_io/cloud is not truly working as it depends on
ReaderBase which seems to be deprecated in tensorflow's core rep.

This PR removes BUILD file tensorflow_io/cloud so that
it will not be build and picked up by wheel package

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Pretty sure gcs_config_ops has been fully covered in tensorflow_io/gcs,

so remove the .cc file as well.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>