Add PBM(PPM/PGM) image support
This PR intends to address the issue raised in tensorflow/tensorflow 31402
where there were no PBM(PPM/PGM) image support.

PBM is not an efficient image format, though it is used in some places,
most notably NYU dataset which consists of RGB+D sequences.

This PR adds the PBM(PPM/PGM) support.

Note as PBM(PPM/PGM) could be either 256 or 65536 so either upscaling or
downscaling might be needed depending on desired output type.

There are a couple of edge combinations that are not covered, though
they could be easily added if test image is available.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>