Enable SSE4.2 and AVX (not AVX2) for tensorflow-io
This PR is part of the effort for 612 and 611. Basically
we enables SSE4.2 and AVX support by default on Linux.

If machines used does not have the SSE4.2 and AVX support
then it will thrown out an error on Linux.

Use could still build their own binary with any instruction set.

Note SSE4.2 and AVX are enabled by tensorflow's core library
already, so we are basically inline with tensorflow.

Note also that AVX2 are not enabled yet, also inline with tensorflow.

This PR fixes 612.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>