Add graph mode support for audio IOTensor and IODataset (#615)
* Add graph mode support for audio IOTensor and IODataset

This PR tries to address the issue in 581 where IOTensor
from audio (WAV) file does not support graph mode, thus
could not be wired up inside tf.data.Dataset.map() function.

There were several reasons:
1) Initially tf.data was designed as a pipeline in graph mode
   with efficiency as the goal.
2) It was not possible to random access with tf.data as
   tf.data always iterate from beginning to end.
3) It was not possible to get meta data as graph mode essentially
   is a one-shot operation and all of the information about shape
   and dtype has to be defined exactly before hand.

Some of the goals of tensorflow/io were to address 2) and 3).
While addressing 3), IOTensor utilize eager mode to obtain the metadata
in one run, then run the graph.

This reduces user knobs and there is no need to specify Tensor
shape and dtype for any operations.

However, since it utilize eager mode, it is not possible to use IOTensor
inside a tf.data.Dataset's `map` function, because `map` will wrap the
function (or lambda) as a subgraph with autograph.

This PR tries to address the issue by avoid eager mode in IOTensor.

However, by nature of graph, at the minimium the dtype of the Tensor
has to be specified for the output node. There is no real way to
get around (other than infered or casted).

The solution in this PR, is to introduce the graph mode and requiring
user to provide the dtype. With dtype in place, this PR also did some
additional plumbing so that AudioGraphIOTensor and AudioGraphIODataset
could be used in graph mode, thus they could be used in combination
with tf.data.Dataset's map function (and tf.data pipeline).

The loss is the need to provide dtype before hand. But user still have
the option of not providing dtype and use AudioIOTensor and AudioIODataset
in eager mode.

This PR fixes 581.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Add `__len__` to expose len(IOTensor) function

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>