Add Kafka message key support
This PR is an enhancement for KafkaDataset to add
an option of `message_key=<False|True>`. By default
`message_key=False` which maintains the previous behavior.

When `message_key=True` is passed, the KafkaDataset will
output a pair of `message, key`. This could be useful
when the key of the message might be needed.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>