Add option to initialize decode_avro resource, so that compiled avro schema could be reused
This PR add a decode_avro_init which returns a resource handle
with compiled avro schema.

The resource handle could then be passed to decode_avro multiple times,
thus avoiding recompile of avro schema.

Note decode_avro could accept both a string (schema) or a resource handle.
If a string is passed, the recompile of schema will be done in place
within the kernel ops of decode_avro. This PR is backward-compatible.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>