[tensorflow-io] Fix memory leak
This commit fixes the memory leak in tensorflow-io avro module.

The memory leak happened due to not having the virtual destructor defined in the
base class. When the derived class gets out of scope, the destructor will not be
called. The memory hold by the derived class will not be free.

RB=1878399
BUG=APA-9517
G=tony-reviewers,tensorflow-reviewers
R=fraudies,htian,xhzhang,chren
A=fraudies,anyan,spopuri,ali