Add `alwayslink = 1` for Bazel rules to support bazel 1.0+
For some reason, Bazel 1.0+ changed the way cc_library is generated,
as a result, with bazel 1.0+, building tensorflow-io fails.

This PR adds `alwayslink = 1` so that code base of tensorflow-io
still works with bazel 1.0+.

Note on Travis CI we still use 0.29.0 for now (as TF 2.1 release is comming).

We will upgrade Travis CI's Bazel version later.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>