Add initial benchmark tests, and add initial standarized IODataset tests
This PR is an effort to:
1) Add benchmark tests
2) Add standarized IODataset tests
so that all IODataset (especially graph mode GraphIODataset)
behaves the same across the board.

At the moment only from_audio has been moved to this standarized
tests with the help of fixtures and parametrized tests in pytest.

This PR also found out that from_lmdb can not repeat the iteration,
in other words:
```
dataset = tfio.IODataset.from_lmdb(...)
entries = [e for e in dataset]
entries = [e for e in dataset] <= failure!
```

This PR has not fixed the lmdb yet, will have a follow up PR
to fix lmdb.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>