Fix string issue in scalar hdf5 type
This PR is a follow up to 708 and 174.

In PR 708, the scalar hdf5 type has been supported.

However, it looks like the variable length string has not been covered
yet. Only fixed length string (e.g, '<S5') are covered.

This PR adds the support to cover the variable length string.

(Note variable length data array has already been covered before
but this is a different type).

This PR fixes 174 and is a follow up to 708.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>