Add structured support for prometheus dataset
This PR add structured support for prometheus dataset,
so that it is possible to obtain mulitple/structured values from
prometheus servers.

The spec of the value structure is:
```
{
  "job_1": {
    "instance_1": {
      "name_1": Tensor,
      "name_2": Tensor,
    },
    "instance_2": {
      "name_1": Tensor,
      "name_2": Tensor,
    },
  },
  "job_2": {
    "instance_1": {
      "name_1": Tensor,
      "name_2": Tensor,
    },
    "instance_2": {
      "name_1": Tensor,
      "name_2": Tensor,
    },
  },
}
```

Note in eager mode, the strucutre of the value is automatically
probed. In graph mode, it has to be provided through the `spec`
to args.

This PR fixes 717.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>