Add PostgreSQL server support for Dataset
This PR adds PostgreSQL server support for Dataset
so that it is possible to query postgresql server with
sql and generate a Dataset.

At the moment only int32/int64/float32/float64 are supported
though other data types could be added later.

The implementation is also not very performance-centric,
as it takes one record at a time. Though would prefer to defer
the optimization in follow up PRs.

We need a pipeline that works first (and on all platforms) as
SQL database interaction could be complicated.

The test is only done on Linux platform, theough build for
Windows and macOS has already been in. Only need to test properly
before states Windows and macOS is supported.

This PR is part of 548.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>