Add GitHub CI to test build instructions on macOS
This PR adds GitHub CI to test build instructions on macOS,
so that the content in README.md's development sections are verified.

This helps us in improving the documentation on README.md so that we know
the README.md is correct in steps.

Note this one is different from the steps we take in building the wheel binary
(building wheel use optimized gcc compiler flags for performance, while
the README.md only guaranteed that build instruction with minimal configuration
works correctly).

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>