Add GeoTIFF support in decode_tiff
This PR essentially replaces PR 578. GeoTIFF is a superset of
TIFF that contains geographic information.

I still haven't figured out exactly the best way to represent
the geographic information so PR 578 has not be moving forward since then.

But, since GeoTIFF is a superset of TIFF, at the minimium we could
replace TIFF internal implementation with GeoTIFF internal information,
so that it is possible to call decode_tiff to process GeoTIFF
to get the image.

This PR internally adds GeoTIFF support while no external API has been changed.

This PR closes 578.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>