Fix hdf5 issue when resource tensor might be reused in graph incorrectly
This PR fixes 841

The reason for the issue was that the resource tensor might be reused in graph incorrectly
in tensorflow code base (likely some change in behavior on how the resource should be intepeareted).

This PR assign a unique container and name to make sure resource created for hdf5 is never re-used.

(The same issue may happen in other ops as well?)

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>