Remove BigTable from the code tree
BigTable has been broken from the beginning, and was never maintained.
In addition, there are lots of code duplication as it implment 6 datasets
in C++ with very minimal difference. Due to the recent changes in TF upsteram
API change, it gets harder and harder to even make it compile.

This PR removes BigTable from the code tree. The plan is to add BigTable in the future
with right implementation, ideally utilize the better maintained google cloud C++ big table API:

https://github.com/googleapis/google-cloud-cpp/tree/master/google/cloud/bigtable seems

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>