Fix Docker file build failure
This PR tries to fix issue raised in 987.

The issue seems to be related to Ubuntu 16.04 have a gcc 5.4 that is old
and does not plays well with std=c++14.
This PR adjust arrow's flag to std=c++11 to allow gcc 5.4 working again.

As python 3.5 will EOL in 09/13/2020, we might consider dropping Ubuntu 16.04
support (and issues related to gcc 5.4 + python 3.5) sometime this year.

This PR fixes 987.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>