support partially-known and unknown shape specification in decode_json
This PR tries to address the isseu in 918 by adding
support for partially-known and unknown shape specification in decode_json.

In this PR, the shapes are not passed in C++ kernels anymore. Instead,
the C++ kernels will always render 1-D output tensor first with a
follow-up tf.reshape to fixup the shape later.

This PR fixes 918.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>