Add support of multiple dimension tensor for tfio.experimental.serialization.decode_json
This PR tries to address the issue raised in 695 where
it was not possible to use tfio.experimental.serialization.decode_json
for entries of multi dimensional tensor, e.g. `'{"x":[[1.0]]}'`.

Supporting unlimited dimensional tensor could be a challange so this PR
takes a first step of support 2D tensor first.

A follow up PR will be created for higher dimension tensors.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>